fix for tracker item 2723518: resampling in 3.98 and 3.99alpha

+++ libmp3lame/util.h
@@ -269,7 +269,7 @@ extern  "C" {
         /* variables used by util.c */
         /* BPC = maximum number of filter convolution windows to precompute */
 #define BPC 320
-        FLOAT   itime[2];
+        double  itime[2]; /* float precision seems to be not enough */
         sample_t *inbuf_old[2];
         sample_t *blackfilt[2 * BPC + 1];
